requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.4.1
Django==1.10.3
django-cors-headers==1.3.1
django-filter==1.0.1
djangorestframework==3.5.3
gunicorn==19.6.0
idna==2.6
jsonpickle==0.9.5
Markdown==2.6.7
matplotlib==1.5.3
numpy==1.11.2
pyparsing==2.1.10
python-dateutil==2.6.0
python-decouple==3.0
pytz==2016.7
requests==2.18.4
six==1.10.0
urllib3==1.22
virtualenv==15.1.0
whitenoise==3.2.2
xlrd==1.0.0
xlwt==1.1.2
-e ./espnFF/
